Renamed vfx_sampling.py file to intermediate_sampling.py to prepare addition of other intermediate samplers
